backend/python/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
python-Levenshtein==0.23.0
fuzzywuzzy==0.18.0
mysql-connector-python==8.2.0
openai==1.3.0
textblob==0.17.1
nltk==3.8.1
